+++ src/frontend/environments/app-environment.ts
@@ -10,4 +10,5 @@ export interface AppEnvironment extends CommonEnvironment {
     tdocWritable: boolean;
     tdocActionTagWritable: boolean;
     tdocMaxItemsPerAlbum: number;
+    m3uAvailable: boolean;
 }